Fix option parsing, add --nozeros option
A custom option parser was added to terminal-interface.sh . Previously,
the BASH getopts built-in was used, but the getopts built-in could not
parse long-form (e.g. --help) options. Additionally, a --nozeros
option was added to terminal-interface.sh, allowing the user to decide
whether to map genes with zero transcription values. However, the actual
capability to not map zero-valued genes has not yet been added to
heatmap-generator.r .