Change alias progression
Previously, the command to run genetic-heatmaps was simply
"heatmap ...". This worked because generating heatmaps was the only
operation of the genetic-heatmaps program.

Currently, "heatmap" is the only operation that is fully implemented,
but more data analysis operations are being developed.

Now that more operations are being added to genetic-heatmaps, a
particular operation is run by typing "ghmtools OPERATION ...".