Add multi-operation help interface
Previously, the command to display a help message was "heatmap --help".
This worked because generating heatmaps was the only operation of the
genetic-heatmaps program and there was only one help message.

Now that more operations are being added to genetic-heatmaps, a
particular operation's help message is displayed by typing
"ghmtools help OPERATION ..." or "ghmtools --help OPERATION".
Additionally, a list of all operations is displayed by typing
"ghmtools help" or "ghmtools --help".